+++ lib/models/stat.dart
@@ -1,8 +1,15 @@
+// -----------------------------------------------------------------------------
+//
+// Stats are single values representing a certain aspect of a being,
+// such as its current health points, speed etc. There is no upper limit.
+//
+// -----------------------------------------------------------------------------
+
 enum StatType {
-  health,
-  mana,
-  focus,
-  xp,  // fills up from 0 - n, and wenn full, the player gains a level
+  health, // health points
+  mana,   // needed to perform magic
+  focus,  // needed to perform special melee attacks / recharges automatically
+  xp,     // fills up from 0 - n, and wenn full, the player gains a level
 }
 
 /**
@@ -12,20 +19,20 @@ enum StatType {
 class Stat {
 
   StatType statType;
-  int _value = 0;
-  int _maxValue;
+  double _value = 0;
+  double _maxValue;
 
-  Stat(StatType statType, int maxValue) : statType = statType, _maxValue = maxValue {
+  Stat(StatType statType, double maxValue) : statType = statType, _maxValue = maxValue {
   }
 
-  Stat.withValue(StatType statType, int value, int maxValue) : statType = statType, _value = value, _maxValue = maxValue {
+  Stat.withValue(StatType statType, double value, double maxValue) : statType = statType, _value = value, _maxValue = maxValue {
   }
 
-  int maxValue() {
+  double maxValue() {
     return _maxValue;
   }
 
-  int value() {
+  double value() {
 //    print("...../ stat: " + statType.name + " / value: " + _value.toString());
     return _value;
   }
@@ -41,11 +48,11 @@ class Stat {
     return 1.0 / (_maxValue.toDouble() / _value.toDouble());
   }
 
-  void setValueTo(int newValue) {
+  void setValueTo(double newValue) {
     _value = newValue;
   }
 
-  void setMaxValueTo(int newMaxValue) {
+  void setMaxValueTo(double newMaxValue) {
     _maxValue = newMaxValue;
   }
 
@@ -53,16 +60,16 @@ class Stat {
     _value = 0;
   }
 
-  void decreaseBy(int value) {
+  void decreaseBy(double value) {
     _value -= value;
     _value = _value < 0 ? 0 : _value;
   }
 
-  void increaseMaxValueBy(int increase) {
+  void increaseMaxValueBy(double increase) {
     _maxValue += increase;
   }
 
-  void restoreBy(int restoreValue) {
+  void restoreBy(double restoreValue) {
     _value += restoreValue;
     _value = _value > _maxValue ? _maxValue : _value;
   }
